refactor(test-utils): move buildParticipants to utils file
