+++ client/test/test-utils.js
@@ -22,5 +22,15 @@ export function buildParticipant(options) {
   };
 }
 
+export function buildParticipants({ maxNumber = 10, number } = {}) {
+  const numberOfParticipants = number ?? Math.ceil(Math.random() * maxNumber);
+
+  let participants = [];
+  for (let i = 0; i < numberOfParticipants; i++) {
+    participants.push(buildParticipant());
+  }
+  return participants;
+}
+
 export * from '@testing-library/react';
 export { render };
